fixed #7752: Allow entering the values with a fractional part

+++ src/projectscene/view/clipsview/cliplistitem.h
@@ -30,7 +30,7 @@ class ClipListItem : public QObject
     Q_PROPERTY(bool selected READ selected WRITE setSelected NOTIFY selectedChanged FINAL)
 
     Q_PROPERTY(int pitch READ pitch NOTIFY pitchChanged FINAL)
-    Q_PROPERTY(int speedPercentage READ speedPercentage NOTIFY speedPercentageChanged FINAL)
+    Q_PROPERTY(double speedPercentage READ speedPercentage NOTIFY speedPercentageChanged FINAL)
 
 public:
     ClipListItem(QObject* parent);
@@ -63,7 +63,7 @@ public:
     void setRightVisibleMargin(double newRightVisibleMargin);
 
     int pitch() const;
-    int speedPercentage() const;
+    double speedPercentage() const;
 
 signals:
     void titleChanged();
